[Fix]経験値が正しく読み込みされない不具合を修正

+++ src/main/scala/com/ryoserver/Player/LoadAllPlayerData.scala
@@ -14,7 +14,7 @@ class LoadAllPlayerData() {
     while (rs.next()) {
       val uuid = UUID.fromString(rs.getString("UUID"))
       val level = rs.getInt("Level")
-      val exp = rs.getInt("EXP")
+      val exp = rs.getDouble("EXP")
       val lastDistributionReceived = rs.getInt("lastDistributionReceived")
       val skillPoint = rs.getDouble("SkillPoint")
       val loginNumber = rs.getInt("loginDays")
